agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic-ai[examples, logfire]>=0.0.52",
    "fastapi[standard]>=0.115.12",
    "pytest>=8.3.5",
    "python-dotenv>=1.1.0",
    "ruff>=0.11.4",
    "streamlit>=1.44.1",
    "uvicorn>=0.34.0",
]

[tool.uv]
dev-dependencies = [
    "httpx>=0.28.1",
    "pytest>=8.3.5",
    "ruff>=0.11.2",
]

[tool.ruff.lint]
extend-select = ["E", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
addopts = '-q'